Refactor parser code into separate file
