+++ src/index.ts
@@ -1,34 +1,12 @@
-import Ajv from "ajv";
 import fs from "fs/promises";
 import glob from "globby";
-import yaml from "js-yaml";
 import path from "path";
 
-import type { Block } from "./block";
-import { blockSchema } from "./block";
-import { matchRuleSchema } from "./matchRule";
-import { satisfyRuleSchema } from "./satisfyRule";
+import { parseYAML } from "./parser";
 import { Verifier } from "./verifier";
 
 const BLOCK_CLASSES = ["primary", "second", "minor"];
 const PATH_PREFIX = path.join(__dirname, "../blocks");
-
-const ajv = new Ajv({
-  allowUnionTypes: true,
-  schemas: [matchRuleSchema, satisfyRuleSchema],
-});
-const ajvValidate = ajv.compile(blockSchema);
-
-const parseYAML = (contents: string): Block => {
-  const block = yaml.load(contents);
-  if (typeof block !== "object" || !block) {
-    throw new Error(`unexpected type: ${typeof block}`);
-  }
-  if (!ajvValidate(block)) {
-    throw new Error(JSON.stringify(ajvValidate.errors, null, 2));
-  }
-  return block;
-};
 
 const loadBlocks = async (type: string): Promise<Verifier> => {
   const files = await glob(`${PATH_PREFIX}/${type}/**/*.yml`);
@@ -53,4 +31,4 @@ const initVerifiers = async (): Promise<Record<string, Verifier>> => {
   );
 };
 
-export { BLOCK_CLASSES, initVerifiers };
+export { BLOCK_CLASSES, initVerifiers, parseYAML };
